+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## $v
+
+- Enhancing security.
+- Adding options page.
+- Adding ability to whitelist atts/values.
+- Adding many new attribute/value possibilities.
+- Refactoring source code to make it more robust.
+
 ## v170308.52831
 
 - Initial release.